requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
pymysql==1.1.1

